requirements.txt
pandas
streamlit
Pillow
bokeh
matplotlib
seaborn
graphviz
plotly
pybase64
